pyproject.toml
[project]
name = "telepromp-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyqt6>=6.4.0",
    "pyside6>=6.10.2",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
]
